requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
Pillow
requests
openai==1.1.1
opencv-python==4.9.0.80
pandas==2.2.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
instructor==0.6.7
python-dotenv==1.0.1
python-multipart==0.0.9
gunicorn==21.2.0
azure-ai-formrecognizer==3.3.0
azure-ai-documentintelligence
